Updating to support GEOS 3.9 required changes to GetWkts(), which is used by BuildAdminFromPBF. Laid down a simple unit test for BuildAdminFromPBF().
