+++ valhalla/mjolnir/adminbuilder.h
@@ -1,0 +1,18 @@
+#pragma once
+
+#include <boost/property_tree/ptree.hpp>
+
+#include <cstdint>
+#include <iostream>
+#include <list>
+#include <string>
+#include <vector>
+
+namespace valhalla {
+namespace mjolnir {
+
+void BuildAdminFromPBF(const boost::property_tree::ptree& pt,
+                       const std::vector<std::string>& input_files);
+
+}
+}
